# api/openapi.yml
openapi: 3.0.3
info:
  title: RASP Service API
  description: API для управления расписанием занятий, бронированиями и посещаемостью
  version: "1.0.0"
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - REQUEST_FAILED
                - FAILED_TO_DECODE
                - NOT_FOUND
                - LOCKED
                - CONFLICT
                - SLOT_NOT_AVAILABLE
            message:
              type: string
      example:
        error:
          code: NOT_FOUND
          message: resource not found
...
    SlotResponse:
      type: object
      required:
        - id
        - start
        - end
        - teacher_id
        - status
      properties:
        id:
          type: string
          description: Уникальный идентификатор слота
        start:
          type: string
          format: date-time
          description: Время начала слота
        end:
          type: string
          format: date-time
          description: Время окончания слота
        teacher_id:
          type: string
          description: Идентификатор преподавателя
        status:
          type: string
          enum:
            - free
            - booked
            - cancelled
            - blocked
          description: Статус слота
        booking_id:
          type: string
          nullable: true
          description: Идентификатор бронирования (если слот забронирован)
        template_id:
          type: string
          nullable: true
          description: Идентификатор шаблона, из которого создан слот
...
  parameters:
...
    IdsQuery:
      name: ids
      in: query
      required: false
      schema:
        type: string
      description: Список идентификаторов через запятую
...
paths:
...
  /slots/batch:
    get:
      tags:
        - Slots
      summary: Получить слоты по идентификаторам
      description: Возвращает список слотов по переданным идентификаторам
      parameters:
        - $ref: '#/components/parameters/IdsQuery'
      responses:
        '200':
          description: Список слотов
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/SlotResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to get slots
...